backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bleach>=6.3.0",
    "gunicorn>=23.0.0",
    "fastapi>=0.121.2",
    "firebase-admin>=7.1.0",
    "pydantic>=2.12.4",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-multipart",
    "slowapi>=0.1.9",
    "uvicorn>=0.38.0",
    "pyotp>=2.9.0",
    "qrcode>=8.0",
    "pillow>=10.0.0",
    "pandas>=2.0.0",
    "google-generativeai",
    "ofxparse",
    "fastapi-mail>=1.6.1",
    "stripe>=14.1.0",
]
